Renamed drivers-license-category
