+++ backend/src/drivers-license-category/drivers-license-category.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { DriversLicenseCategoriesController } from './driversLicenseCategory.controller';
-import { DriversLicenseCategoriesService } from './driversLicenseCategory.service';
-import { DriversLicenseCategory } from './entities/driversLicenseCategory.entity';
+import { DriversLicenseCategoriesController } from './drivers-license-category.controller';
+import { DriversLicenseCategoriesService } from './drivers-license-category.service';
+import { DriversLicenseCategory } from './entities/drivers-license-category.entity';
 
 @Module({
   controllers: [DriversLicenseCategoriesController],
